Refactoring dynamo db code into util file
